done v1.0

+++ wp-content/plugins/request-a-brochure/inc/api/rab-brochure-controller.php
@@ -10,6 +10,10 @@ class RAB_Brochure_Controller
 
     public function get_brochures()
     {
+        if (!is_admin()) {
+            return new WP_REST_Response([], 401);
+        }
+
         $brochures = $this->rs->get_all_brochures();
         if (!$brochures) {
             return new WP_REST_Response([], 404);
@@ -20,16 +24,22 @@ class RAB_Brochure_Controller
 
     public function create_brochure(string $brochure)
     {
+        if (!is_admin()) {
+            return new WP_REST_Response([], 401);
+        }
         $brochure = $this->rs->create_brochure($brochure);
         if (!$brochure) {
             return new WP_REST_Response($brochure, 500);
         }
 
-        return new WP_REST_Response($brochure,200);
+        return new WP_REST_Response($brochure, 200);
     }
 
     public function delete_brochure(int $id)
     {
+        if (!is_admin()) {
+            return new WP_REST_Response([], 401);
+        }
         $res = $this->rs->delete_brochure($id);
         if (!$res) {
             return new WP_REST_Response($res, 500);
@@ -39,6 +49,9 @@ class RAB_Brochure_Controller
 
     public function update_brochure_status(int $id, int $status)
     {
+        if (!is_admin()) {
+            return new WP_REST_Response([], 401);
+        }
         $res = $this->rs->update_brochure_status($id, $status);
         if (!$res) {
             return new WP_REST_Response($res, 500);
